Separate cargo and build caches
The job that runs dev checks (fmt, clippy, test) only uses and generates
code in target/debug. Similarly, the job that builds the docker image
only accesses target/release. Because of that they would interfere with
each other by uploading/downloading a version of the cache where only
"their own" directory was populated and that was useless for the other
job. By making this separation, each job keeps a separate, smaller cache
of only the things that it's interested in.