Tweak GHA caching (#32)
* Tweak GHA caching

* Cache whole ~/.cargo in GHA

... except for ~/.cargo/registry/src, which is quite large. The goal
here is to cache packages installed with `cargo install`.

* Cache the same way in docker and check jobs

* Separate cargo and build caches

The job that runs dev checks (fmt, clippy, test) only uses and generates
code in target/debug. Similarly, the job that builds the docker image
only accesses target/release. Because of that they would interfere with
each other by uploading/downloading a version of the cache where only
"their own" directory was populated and that was useless for the other
job. By making this separation, each job keeps a separate, smaller cache
of only the things that it's interested in.

* Extract caching into separate action

* Unify actions/cache version

* Fix typo

* Rename checks -> rust-checks

* Change default target_key to ''