Extract scenario ident/interval to a common place
This should make the scenarios themselves cleaner to write and maintain.
The downside is that now you have to thread `logger` and `connection`
through the calls in the scenario... I think this is worth it, but I may
be skewed - I'm used to functional programming, where you often pass a
bunch of stuff like that around.